BAU: Restructure app
