--- requirements.txt
@@ -1,12 +1,0 @@
-boto3==1.16.50
-botocore==1.19.50
-colorama==0.4.4
-commonmark==0.9.1
-jmespath==0.10.0
-Pygments==2.7.3
-python-dateutil==2.8.1
-rich==9.6.1
-s3transfer==0.3.3
-six==1.15.0
-typing-extensions==3.7.4.3
-urllib3==1.26.2
